+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.2.16] - 2026-07-23
+
 ### Added
 
 - ZFS checks for linux nodes: Orbit now parses the Checkmk